# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Notification API",
    "description": "An API for managing user notifications",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get all notifications",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notification"
              }
            }
          }
        }
      }
    },
    "/notifications/unread": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get all unread notifications",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notification"
              }
            }
          }
        }
      }
    },
    "/notifications/unread/count": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get the count of unread notifications",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "/notifications/read": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get all read notifications",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notification"
              }
            }
          }
        }
      }
    },
    "/notifications/read/count": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get the count of read notifications",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "/notifications/{id}": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get a notification by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the notification to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Notification"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark a notification as read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the notification to mark as read",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/notifications/{id}/unread": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark a notification as unread",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the notification to mark as unread",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/notifications/user/{userId}": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get all notifications for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user to get notifications for",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notification"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Notification": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "read": {
          "type": "boolean"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    }
  }
}